fix: address encoding based on family
