+++ deployment/ccip/view/shared/common.go
@@ -2,6 +2,8 @@ package shared
 
 import (
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/mr-tron/base58"
+	chain_selectors "github.com/smartcontractkit/chain-selectors"
 
 	"github.com/smartcontractkit/chainlink-ccip/chains/evm/gobindings/generated/v1_5_0/token_admin_registry"
 )
@@ -25,4 +27,20 @@ func GetSupportedTokens(taContract *token_admin_registry.TokenAdminRegistry) ([]
 		}
 	}
 	return allTokens, nil
+}
+
+func GetAddressFromBytes(chainSelector uint64, address []byte) string {
+	family, err := chain_selectors.GetSelectorFamily(chainSelector)
+	if err != nil {
+		return "invalid chain selector"
+	}
+
+	switch family {
+	case chain_selectors.FamilyEVM:
+		return common.BytesToAddress(address).Hex()
+	case chain_selectors.FamilySolana:
+		return base58.Encode(address)
+	default:
+		return "unsupported chain family"
+	}
 }
